airflow-delta/requirements.txt
apache-airflow-providers-apache-spark==4.9.0
apache-airflow-providers-apache-kafka==1.6.1
pyarrow==18.0.0
pyspark==3.5.1
delta-spark==3.2.0
pymysql==1.1.1
xmltodict==0.14.2